Cleaned up landing, login and register routes
