+++ views/index.html
@@ -7,7 +7,7 @@
 
     <body>
 
-        <form action="/createuser" method="post">
+        <form action="/register/createuser" method="post">
             <input name="email" type="email"/>
             <input name="user" type="text"/>
             <input name="pass" type="password"/>
