+++ views/pages/register.ejs
@@ -21,8 +21,8 @@
                 <div class="floating-box__login__icon">
                     <img class="vertC__relative" src="/img/icons/login.svg"/>
                 </div>
-                <h3>Admin Login</h3>
-                <form action="/login/sign-in" method="post">
+                <h3>Register</h3>
+                <form action="/register/create-user" method="post">
                     <input type="text" name="user" placeholder="Username"/>
                     <input type="password" name="pass" placeholder="Password"/>
                     <input type="submit" placeholder="Login"/>
